+++ tests/grammar.test.ts
@@ -1,6 +1,7 @@
 import { expect, test } from "vitest";
-import { AbstractGrammarFactory  } from "../src/engine/types/grammar_factories.ts";
-import { AddProductionRuleCommand, GrammarType } from "../src/engine/types/grammar_types.ts";
+import { AbstractGrammarFactory } from "../src/engine/grammar/factories.ts";
+import { AddProductionRuleCommand } from "../src/engine/grammar/commands/edit.ts";
+import { GrammarType } from "../src/engine/grammar/grammar.ts";
 
 test("Regular grammar test", () => {
   const agf = new AbstractGrammarFactory(GrammarType.REGULAR);
